requirements.txt
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
numpy==1.26.4
tqdm==4.67.1
Pillow==11.2.1
packaging==25.0
transformers==4.41.0
qwen-vl-utils==0.0.8
huggingface-hub==0.30.2
tokenizers==0.19.1
safetensors==0.5.3
einops==0.8.1
timm==1.0.15
xformers==0.0.29.post2
flash-attn==2.8.0.post2
bitsandbytes==0.46.0
opencv-python==4.11.0.86
scipy==1.15.2
requests==2.32.3
filelock==3.18.0
regex==2024.11.6
